Initial import from the latest DRUPAL-6--4 revision (2009-Oct-19).
No history has been imported since the cvs history of this project
has some mess in it. cvsimport was unable to correctly import the
branches and I was unable to fix the git repo created.

The original DRUPAL-7 branch was based on and old version and the
theme has been overhault since then. So a new d7 branch will be
created based on DRUPAL-6-4.